Started work on feature17. Js-files refactoring, header added for all present pages.
