+++ resources/js/login.js
@@ -1,3 +1,7 @@
+import {ThrowError} from "./throwError.js";
+import {sendRequest} from "./sendRequest.js";
+
+
 document.addEventListener('DOMContentLoaded', () => {
     //Методы отправки данных для авторизации
     let _button_signup = document.getElementById("signup-button");
@@ -29,31 +33,6 @@ document.addEventListener('DOMContentLoaded', () => {
     function RedirectFromLogin(url, cookie_value) {
         addCookie('Authenticate', cookie_value)
         window.location.replace(url)
-    }
-
-    function ThrowError(data) {
-        console.log(data)
-    }
-
-
-    function sendRequest(method, url, json) {
-        return new Promise((resolve, reject) => {
-            const xhr = new XMLHttpRequest()
-            xhr.open(method, url)
-            xhr.responseType = 'json'
-            xhr.setRequestHeader('Content-type', 'application/json')
-            xhr.onload = () => {
-                if (xhr.status >= 400) {
-                    reject(xhr.response)
-                } else {
-                    resolve(xhr.response)
-                }
-            }
-            xhr.onerror = () => {
-                console.log(onerror)
-            }
-            xhr.send(JSON.stringify(json))
-        })
     }
 
     function addCookie(name, value) {
