--- resources/js/pieces.js
@@ -1,37 +1,0 @@
-document.addEventListener('DOMContentLoaded', () => {
-    // let headerInner = `
-    // <div class="container">
-    //     <div class="d-flex flex-wrap align-items-center justify-content-center justify-content-lg-start">
-    //         <a href="/" class="d-flex align-items-center mb-2 mb-lg-0 text-white text-decoration-none">
-    //             <img width="32" height="32" class="mx-3" src="../icons/logo_white.png" alt="logo"/>
-    //         </a>
-    //
-    //         <ul class="nav col-12 col-lg-auto me-lg-auto mb-2 justify-content-center mb-md-0">
-    //             <li><a href="../templates/main.html" class="nav-link px-2 text-secondary">Home</a></li>
-    //             <li><a href="../templates/" class="nav-link px-2 text-white">Features</a></li>
-    //             <li><a href="#" class="nav-link px-2 text-white">Pricing</a></li>
-    //             <li><a href="#" class="nav-link px-2 text-white">FAQs</a></li>
-    //             <li><a href="#" class="nav-link px-2 text-white">About</a></li>
-    //         </ul>
-    //
-    //         <form class="col-12 col-lg-auto mb-3 mb-lg-0 me-lg-3">
-    //             <input type="search" class="form-control form-control-dark" placeholder="Search..." aria-label="Search">
-    //         </form>
-    //
-    //         <div class="text-end">
-    //             <a href="../templates/login.html" type="button" class="btn me-2">Login</a>
-    //             <a href="../templates/login.html" type="button" class="btn me-2">Profile</a>
-    //         </div>
-    //     </div>
-    // </div>
-    // `
-    //
-    // function setHeader() {
-    //     let header = document.createElement("header");
-    //     header.setAttribute("class", "p-3 bg-dark text-white");
-    //     header.innerHTML = headerInner;
-    //     document.body.insertAdjacentElement('afterbegin', header);
-    // }
-    //
-    // setHeader();
-})
