+++ resources/js/sendRequest.js
@@ -1,0 +1,19 @@
+export function sendRequest(method, url, json) {
+    return new Promise((resolve, reject) => {
+        const xhr = new XMLHttpRequest()
+        xhr.open(method, url)
+        xhr.responseType = 'json'
+        xhr.setRequestHeader('Content-type', 'application/json')
+        xhr.onload = () => {
+            if (xhr.status >= 400) {
+                reject(xhr.response)
+            } else {
+                resolve(xhr.response)
+            }
+        }
+        xhr.onerror = () => {
+            console.log(onerror)
+        }
+        xhr.send(JSON.stringify(json))
+    })
+}
