+++ resources/js/throwError.js
@@ -1,0 +1,3 @@
+export function ThrowError(data) {
+    console.log(data)
+}
