+++ resources/templates/login.html
@@ -79,6 +79,6 @@
         </div>
     </div>
 </div>
-<script src="../js/login.js"></script>
+<script type="module" src="../js/login.js"></script>
 </body>
 </html>
