+++ resources/templates/profile.html
@@ -17,6 +17,7 @@
             <ul class="nav col-12 col-lg-auto me-lg-auto mb-2 justify-content-center mb-md-0">
                 <li><a href="/" class="nav-link px-2 link-light">Main page</a></li>
                 <li><a href="/teams" class="nav-link px-2 link-light">Teams</a></li>
+                <li><a href="/team" class="nav-link px-2 link-light">My team</a></li>
                 <li><a href="/users" class="nav-link px-2 link-light">Users</a></li>
                 <li><a href="/about" class="nav-link px-2 link-light">About</a></li>
             </ul>
@@ -31,9 +32,8 @@
                     <img src="../icons/logo_white.png" alt="mdo" width="32" height="32" class="rounded-circle">
                 </a>
                 <ul class="dropdown-menu text-small" aria-labelledby="dropdownUser">
-                    <li><a class="dropdown-item" href="#">New team</a></li>
-                    <li><a class="dropdown-item" href="#">Find team/user</a></li>
-                    <li><a class="dropdown-item" href="#">About us</a></li>
+                    <li><a class="dropdown-item" href="/profile">Profile</a></li>
+                    <li><a class="dropdown-item" href="/about">About us</a></li>
                     <li>
                         <hr class="dropdown-divider">
                     </li>
@@ -48,5 +48,6 @@
         integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
         crossorigin="anonymous"></script>
 <script src="../js/profile.js"></script>
+<script type="module" src="../js/main.js"></script>
 </body>
 </html>
